requirements.txt
#Python (3+)
numpy==1.16.0
pandas==0.24.2
praat-parselmouth
moviepy
scikit-image
sk-video
watchtower
opencv-python
webrtcvad
pysptk
imutils
dlib==19.13.0
coloredlogs
h5py
ujson
numba==0.48.0
librosa
more_itertools
scipy==1.2.0
